Reorganize structure. Upload data files. Simply divide data and source files into /data and /src folders.
