+++ data/Melbourne-bbox.kml
@@ -1,0 +1,123 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<kml xmlns='http://www.opengis.net/kml/2.2'>
+	<Document>
+		<name>Melbourne</name>
+		<description><![CDATA[One-day tour?<br>]]></description>
+		<Folder>
+			<name>Melbourne</name>
+			<Placemark>
+				<name>Melbourne bbox</name>
+				<styleUrl>#poly-FF0000-1-128-nodesc</styleUrl>
+				<ExtendedData>
+				</ExtendedData>
+				<Polygon>
+					<outerBoundaryIs>
+						<LinearRing>
+							<tessellate>1</tessellate>
+                            <coordinates>144.597363,-37.591764,0.0
+                                         144.597363,-38.072257,0.0 
+                                         145.360413,-38.072257,0.0 
+                                         145.360413,-37.591764,0.0 
+                                         144.597363,-37.591764,0.0</coordinates>
+						</LinearRing>
+					</outerBoundaryIs>
+				</Polygon>
+			</Placemark>
+			<Placemark>
+				<name>big-box</name>
+				<styleUrl>#poly-93D7E8-8-94-nodesc</styleUrl>
+				<ExtendedData>
+				</ExtendedData>
+				<Polygon>
+					<outerBoundaryIs>
+						<LinearRing>
+							<tessellate>1</tessellate>
+                            <coordinates>141.9,-35.8,0.0 
+                                         141.9,-39.3,0.0 
+                                         147.1,-39.3,0.0 
+                                         147.1,-35.8,0.0
+                                         141.9,-35.8,0.0 
+                            </coordinates>
+						</LinearRing>
+					</outerBoundaryIs>
+				</Polygon>
+			</Placemark>
+		</Folder>
+		<Style id='poly-93D7E8-8-94-nodesc-normal'>
+			<LineStyle>
+				<color>ffE8D793</color>
+				<width>8</width>
+			</LineStyle>
+			<PolyStyle>
+				<color>5EE8D793</color>
+				<fill>1</fill>
+				<outline>1</outline>
+			</PolyStyle>
+			<BalloonStyle>
+				<text><![CDATA[<h3>$[name]</h3>]]></text>
+			</BalloonStyle>
+		</Style>
+		<Style id='poly-93D7E8-8-94-nodesc-highlight'>
+			<LineStyle>
+				<color>ffE8D793</color>
+				<width>12.0</width>
+			</LineStyle>
+			<PolyStyle>
+				<color>5EE8D793</color>
+				<fill>1</fill>
+				<outline>1</outline>
+			</PolyStyle>
+			<BalloonStyle>
+				<text><![CDATA[<h3>$[name]</h3>]]></text>
+			</BalloonStyle>
+		</Style>
+		<StyleMap id='poly-93D7E8-8-94-nodesc'>
+			<Pair>
+				<key>normal</key>
+				<styleUrl>#poly-93D7E8-8-94-nodesc-normal</styleUrl>
+			</Pair>
+			<Pair>
+				<key>highlight</key>
+				<styleUrl>#poly-93D7E8-8-94-nodesc-highlight</styleUrl>
+			</Pair>
+		</StyleMap>
+		<Style id='poly-FF0000-1-128-nodesc-normal'>
+			<LineStyle>
+				<color>ff0000FF</color>
+				<width>1</width>
+			</LineStyle>
+			<PolyStyle>
+				<color>800000FF</color>
+				<fill>1</fill>
+				<outline>1</outline>
+			</PolyStyle>
+			<BalloonStyle>
+				<text><![CDATA[<h3>$[name]</h3>]]></text>
+			</BalloonStyle>
+		</Style>
+		<Style id='poly-FF0000-1-128-nodesc-highlight'>
+			<LineStyle>
+				<color>ff0000FF</color>
+				<width>2.0</width>
+			</LineStyle>
+			<PolyStyle>
+				<color>800000FF</color>
+				<fill>1</fill>
+				<outline>1</outline>
+			</PolyStyle>
+			<BalloonStyle>
+				<text><![CDATA[<h3>$[name]</h3>]]></text>
+			</BalloonStyle>
+		</Style>
+		<StyleMap id='poly-FF0000-1-128-nodesc'>
+			<Pair>
+				<key>normal</key>
+				<styleUrl>#poly-FF0000-1-128-nodesc-normal</styleUrl>
+			</Pair>
+			<Pair>
+				<key>highlight</key>
+				<styleUrl>#poly-FF0000-1-128-nodesc-highlight</styleUrl>
+			</Pair>
+		</StyleMap>
+	</Document>
+</kml>
